requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
databases==0.9.0
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
faiss-cpu==1.11.0
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.44
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.84.0
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
posthog==4.2.0
propcache==0.3.1
protobuf==5.29.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyMySQL==1.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
